Added comments to the code
Added comments to the data classes and the AddEditscenes

+++ src/DataStructure/Data/Performer.java
@@ -2,6 +2,8 @@ package DataStructure.Data;
 
 import java.io.Serializable;
 
+//the performer class is a collection of every object that can be planned in the agenda module, every performer has a name and a populartiy value
+//the popularity is a percentage of how much people will probably go to the concert
 public class Performer implements Serializable {
     protected String performerName;
     private int popularity;
@@ -18,7 +20,7 @@ public class Performer implements Serializable {
     }
 
     public int getPopularity(){
-        int convertedPopularity = (int) (popularity);
+        int convertedPopularity =popularity;
         return convertedPopularity;
     }
